+++ modules/debugger/tests/cases/while_cli.m
@@ -1,0 +1,17 @@
+%=============================================================================
+% Copyright (c) 2016-present Allan CORNET (Nelson)
+%=============================================================================
+% This file is part of Nelson.
+%=============================================================================
+% LICENCE_BLOCK_BEGIN
+% SPDX-License-Identifier: LGPL-3.0-or-later
+% LICENCE_BLOCK_END
+%=============================================================================
+%<--INTERACTIVE TEST-->
+%=============================================================================
+TEST_CONDITION = true;
+while TEST_CONDITION
+    disp('Inside while loop');
+    disp('next line');
+    TEST_CONDITION = true;
+end